simulator/external/habitat-sim-geodesic/pyproject.toml
[tool.isort]
known_third_party = ["cppimport", "numpy", "quaternion"]
multi_line_output = 3
force_grid_wrap = false
line_length = 88
include_trailing_comma = true
use_parentheses = true

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | obselete
    | deps
  )/
)
'''

[tool.poetry]
name = "habitat-sim-geodesic"
version = "0.1.0"
description = ""
authors = ["Erik Wijmans <etw@gatech.edu>"]
include = ["habitat_sim_geodesic/navmeshes/*", "habitat_sim_geodesic/bindings.cpp", "habitat_sim_geodesic/csrc/*"]

[tool.poetry.dependencies]
python = "^3.7"
pybind11 = "^2.4.3"
cppimport = "^18.11.8"
numpy = "^1.18.2"
numpy-quaternion = "^2019.12.11"
numba = "^0.48.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
